requirements.txt
transformers==4.42.3
xtuner[deepspeed]==0.1.23 # only support python==3.10
timm==1.0.9
mmdet==3.3.0
hydra-core==1.3.2
ninja==1.11.1
decord==0.6.0
peft==0.11.1
iopath
pycocoevalcap
scikit-image
opencv-python
tqdm
pillow
numpy
matplotlib
black==24.2.0
usort==1.0.2
ufmt==2.0.0b2
fvcore>=0.1.5.post20221221
tensorboard>=2.17.0
tensordict==0.6.0
submitit>=1.5.1